+++ src/include/device.h
@@ -13,6 +13,14 @@
 #include "rccl_float8.h"
 #include <hip/hip_bfloat16.h>
 #include "nccl_tuner.h"
+#if ROCM_VERSION >= 60000
+   // hip_bf16.h should be used from ROCm 6.0
+  #include <hip/hip_bf16.h>
+  typedef __hip_bfloat16 hip_bfloat16;
+#else
+  #include <hip/hip_bfloat16.h>
+#endif
+#include "nccl_common.h"
 #include "bitops.h"
 #if defined(ENABLE_NPKIT)
 #include "npkit/npkit_struct.h"